+++ TinyNvidiaUpdateChecker/SelectGPU.Designer.cs
@@ -36,10 +36,10 @@
             // 
             // SelectBtn
             // 
-            this.SelectBtn.Location = new System.Drawing.Point(9, 83);
-            this.SelectBtn.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.SelectBtn.Location = new System.Drawing.Point(12, 102);
+            this.SelectBtn.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
             this.SelectBtn.Name = "SelectBtn";
-            this.SelectBtn.Size = new System.Drawing.Size(58, 24);
+            this.SelectBtn.Size = new System.Drawing.Size(77, 30);
             this.SelectBtn.TabIndex = 1;
             this.SelectBtn.Text = "OK";
             this.SelectBtn.UseVisualStyleBackColor = true;
@@ -48,28 +48,29 @@
             // GPUBox
             // 
             this.GPUBox.FormattingEnabled = true;
-            this.GPUBox.Location = new System.Drawing.Point(9, 10);
-            this.GPUBox.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.GPUBox.ItemHeight = 16;
+            this.GPUBox.Location = new System.Drawing.Point(12, 12);
+            this.GPUBox.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
             this.GPUBox.Name = "GPUBox";
-            this.GPUBox.Size = new System.Drawing.Size(187, 69);
+            this.GPUBox.Size = new System.Drawing.Size(248, 84);
             this.GPUBox.TabIndex = 2;
             // 
             // label1
             // 
             this.label1.AutoSize = true;
-            this.label1.Location = new System.Drawing.Point(69, 113);
-            this.label1.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
+            this.label1.Location = new System.Drawing.Point(92, 139);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(125, 13);
+            this.label1.Size = new System.Drawing.Size(161, 17);
             this.label1.TabIndex = 3;
             this.label1.Text = "Select your NVIDIA GPU";
             // 
             // NonRecBox
             // 
             this.NonRecBox.AutoSize = true;
-            this.NonRecBox.Location = new System.Drawing.Point(72, 88);
+            this.NonRecBox.Location = new System.Drawing.Point(96, 108);
+            this.NonRecBox.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.NonRecBox.Name = "NonRecBox";
-            this.NonRecBox.Size = new System.Drawing.Size(124, 17);
+            this.NonRecBox.Size = new System.Drawing.Size(161, 21);
             this.NonRecBox.TabIndex = 4;
             this.NonRecBox.Text = "Show non-supported";
             this.NonRecBox.UseVisualStyleBackColor = true;
@@ -77,15 +78,15 @@
             // 
             // SelectGPU
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(207, 135);
+            this.ClientSize = new System.Drawing.Size(276, 166);
             this.Controls.Add(this.NonRecBox);
             this.Controls.Add(this.label1);
             this.Controls.Add(this.GPUBox);
             this.Controls.Add(this.SelectBtn);
             this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedToolWindow;
-            this.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
             this.Name = "SelectGPU";
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
             this.Text = "TinyNvidiaUpdateChecker - GPU";